configure telescope from telescope module
